Manage sha512256 code as a separate lib
