+++ core/lib/sha512256/alignedarray.h
