+++ core/lib/sha512256/portability.h
